+++ libs/middletier/FileSystemStorageClasses.cpp
@@ -8,7 +8,7 @@
 //------------------------------------------------------------------------------
 void cta::FileSystemStorageClasses::createStorageClass(
   const std::string &name,
-  const uint8_t nbCopies,
+  const uint16_t nbCopies,
   const UserIdentity &creator,
   const std::string &comment) {
   try {
