+++ libs/middletier/MockMiddleTierAdmin.hpp
@@ -100,7 +100,7 @@ public:
   void createStorageClass(
     const SecurityIdentity &requester,
     const std::string &name,
-    const uint8_t nbCopies,
+    const uint16_t nbCopies,
     const std::string &comment);
 
   /**
@@ -177,7 +177,7 @@ public:
   void createArchiveRoute(
     const SecurityIdentity &requester,
     const std::string &storageClassName,
-    const uint8_t copyNb,
+    const uint16_t copyNb,
     const std::string &tapePoolName,
     const std::string &comment);
 
@@ -193,7 +193,7 @@ public:
   void deleteArchiveRoute(
     const SecurityIdentity &requester,
     const std::string &storageClassName,
-    const uint8_t copyNb);
+    const uint16_t copyNb);
 
   /**
    * Gets the current list of archive routes.
